+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/lang/WorldTranslations.kt
@@ -2,6 +2,7 @@ package de.md5lukas.waypoints.lang
 
 import de.md5lukas.commons.paper.registerEvents
 import de.md5lukas.waypoints.events.ConfigReloadEvent
+import java.util.*
 import net.kyori.adventure.text.Component
 import org.bukkit.World
 import org.bukkit.event.EventHandler
@@ -13,18 +14,18 @@ class WorldTranslations(private val tl: YmlTranslationLoader) : Listener {
     tl.plugin.registerEvents(this)
   }
 
-  private val warned: MutableSet<World> = mutableSetOf()
+  private val warned: MutableSet<World> = Collections.synchronizedSet(mutableSetOf())
 
   fun getWorldName(world: World): Component {
     if (world in warned) return Component.text(world.name)
 
     val key = "worlds.${world.name}"
-    return if (key in tl) {
+    return if (tl.contains(key, false)) {
       Component.text(tl[key])
     } else {
       warned.add(world)
       tl.plugin.slF4JLogger.warn(
-          "The world {} has no translation present. Using actual name as a fallback.", world.name)
+          "The world '{}' has no translation present. Using actual name as a fallback.", world.name)
       Component.text(world.name)
     }
   }